libbpf: introduce btfloader
btfloader receives socket & pipe fds as well as the path to a bpf .o
file from bpfloader. It parses the bpf file's BTF information & loads
it into the kernel, then uses the socket & pipe fds to pass back the
BTF fd and the BTF type ids for map keys and values, respectively.

Bug: 203823368
Test: build & boot
Test: libbpf_load_test
Signed-off-by: Connor O'Brien <connoro@google.com>
Change-Id: I37843c2bb16e07ce35c6fc758159db5ed2a85521