btfloader: fix BTF loading error on 5.4 kernel
When BTF info is generated for functions, some of the bits in the info
field of struct btf_type are used to denote static vs. global
vs. extern functions. However, the 5.4 kernel predates this change and
expects these bits to be zero for any function type. As a result, BTF
containing any global function types will fail to load.

To accommodate this, retry with 5.4-friendly BTF if
btf__load_into_kernel fails with EINVAL.

Test: libbpf_load_test passes on 5.4 cuttlefish
Bug: 203823368
Signed-off-by: Connor O'Brien <connoro@google.com>
Change-Id: Ia4841114ef691f89982b9b1a4ad95d7b16197936